first commit
This is my second project, a basic password manager built in Java using Swing for the UI and SQLite for the database. It enables a single user to securely save, view, and delete passwords, with the added ability to generate and validate strong passwords. The login system uses a master key to access stored data. I plan to implement encryption and improve overall security in future updates.